Sredjen entity i DTO

+++ Mail/src/main/java/uns/ac/rs/Mail/dto/RuleDTO.java
@@ -5,24 +5,24 @@ import uns.ac.rs.Mail.entity.Rule;
 import java.io.Serializable;
 
 public class RuleDTO implements Serializable {
-    private int id;
+    private Long id;
     private Rule.Condition condition;
     private Rule.Operation operation;
 
     public RuleDTO() {
     }
 
-    public RuleDTO(int id, Rule.Condition condition, Rule.Operation operation) {
+    public RuleDTO(Long id, Rule.Condition condition, Rule.Operation operation) {
         this.id = id;
         this.condition = condition;
         this.operation = operation;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
